# packages/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona API",
    "summary": "Hosted service for building and running typed-memory AI personas.",
    "version": "0.8.0"
  },
  "paths": {
...
    "/v1/personas/{persona_id}/runs": {
      "post": {
        "tags": ["runs"],
        "summary": "Start Run",
        "description": "Start an agentic run (returns the run_id immediately; runs in background).",
        "operationId": "start_run_v1_personas__persona_id__runs_post",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "tags": ["runs"],
        "summary": "Get Run",
        "description": "Get a run's status + accumulated steps (RLS-scoped \u2192 404).",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "tags": ["runs"],
        "summary": "Stream Events",
        "description": "Stream the run's events as SSE (live, from the in-process event bus).",
        "operationId": "stream_events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/respond": {
      "post": {
        "tags": ["runs"],
        "summary": "Respond",
        "description": "Deliver an answer to a run awaiting an ask-user question.",
        "operationId": "respond_v1_runs__run_id__respond_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToRunRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": ["runs"],
        "summary": "Cancel",
        "description": "Cancel a running run (stops at the next step boundary \u2192 cancelled).",
        "operationId": "cancel_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel V1 Runs  Run Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RespondToRunRequest": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["answer"],
        "title": "RespondToRunRequest",
        "description": "Answer an ask-user question raised by a running agentic loop (\u00a75.3)."
      },
      "RunStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "persona_id": {
            "type": "string",
            "title": "Persona Id"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "persona_id", "task", "status"],
        "title": "RunStatusResponse",
        "description": "A run's status + its accumulated steps (JSON-serialised Run/Step)."
      },
...
      "StartRunRequest": {
        "properties": {
          "task": {
            "type": "string",
            "minLength": 1,
            "title": "Task"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["task"],
        "title": "StartRunRequest",
        "description": "Start an agentic run for a task (\u00a75.3)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}